agents/financial-operations-assistant/requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
xero-python>=6.0.0